Added sudo commands
This fix resolved the issue where NRPE did not have the rights to access the /root/ldap_admin_pass file to properly execute the ldapadd, ldapdelete, and ldapsearch commands. An nrpe_sudoers file will be provided as well as steps to change the nrpe user in /etc/passwd to ensure this plugin works as intended.